add helpful messaging when JMS Provider is not available
When the JMS adapter attempts to create a new connection, it dynamically
imports the JMS Provider and can fail with an opaque `NameError` if no
JMS Provider is on the classpath (or specified with the `require_jars`
directive).

Tests that currently fail due to a JMS Provider being missing are marked
as `pending`, and logstash-plugins/logstash-output-jms#10 has been filed
to track the automation of loading a JMS Provider for testing purposes.